[FIX] 마이페이지 파일 나누기(#76)
