+++ Plu-iOS/Plu-iOS/Scenes/MyPage/Adaptor/MyPageAdaptorDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  MyPageAdaptorDelegate.swift
+//  Plu-iOS
+//
+//  Created by uiskim on 2023/12/24.
+//
+
+import Foundation
+
+protocol MyPageAdaptorDelegate: AnyObject {
+    func isAccept()
+}
