+++ Plu-iOS/Plu-iOS/Scenes/MyPage/MangerImpl/MyPageManagerStub.swift
@@ -1,0 +1,14 @@
+//
+//  MyPageManagerStub.swift
+//  Plu-iOS
+//
+//  Created by uiskim on 2023/12/24.
+//
+
+import Foundation
+
+final class MyPageManagerStub: MyPageManager {
+    func getUserData() async throws -> MyPageUserData {
+        return .dummyData
+    }
+}
