+++ Plu-iOS/Plu-iOS/Scenes/MyPage/Model/MypageOutput.swift
@@ -1,0 +1,14 @@
+//
+//  MypageOutput.swift
+//  Plu-iOS
+//
+//  Created by uiskim on 2023/12/24.
+//
+
+import Foundation
+import Combine
+
+struct MypageOutput {
+    var viewWillAppearPublisher: AnyPublisher<MyPageUserData, Never>
+    var switchOnSubject: PassthroughSubject<Void, Never>
+}
