+++ Plu-iOS/Plu-iOS/Scenes/MyPage/ViewModel/MyPageViewModel.swift
@@ -1,0 +1,10 @@
+//
+//  MyPageViewModel.swift
+//  Plu-iOS
+//
+//  Created by uiskim on 2023/12/24.
+//
+
+import Foundation
+
+protocol MyPageViewModel: ViewModel where Input == MypageInput, Output == MypageOutput {}
